+++ CHANGELOG.md
@@ -12,8 +12,10 @@
   - thread-pool size etc.
 
 
+- [Todo] Drop `bract.ring.wrapper/(after, doafter)` in favour of `bract.core.util/*`
 
-## [WIP] 0.6.2-0.2.0-alpha4 / 2021-February-??
+
+## 0.6.2-0.2.0-alpha4 / 2021-February-14
 
 - Use bract.core 0.6.2-alpha4
 - Log events in wrappers